Move CampaignPreviewCard to shard directory
