+++ js/src/components/paid-ads/campaign-preview/campaign-preview-card.scss
