+++ js/src/setup-ads/ads-stepper/create-campaign/index.js
@@ -13,7 +13,7 @@ import StepContentFooter from '.~/components/stepper/step-content-footer';
 import AppDocumentationLink from '.~/components/app-documentation-link';
 import CreateCampaignFormContent from '.~/components/paid-ads/create-campaign-form-content';
 import AppButton from '.~/components/app-button';
-import CampaignPreviewCard from './campaign-preview-card';
+import { CampaignPreviewCard } from '.~/components/paid-ads/campaign-preview';
 
 /**
  * @fires gla_documentation_link_click with `{ context: 'setup-ads', link_id: 'see-what-ads-look-like', href: 'https://support.google.com/google-ads/answer/6275294' }`
